requirements.txt
fastapi==0.94.1
gradio==3.21.0
imgaug==0.4.0
matplotlib==3.7.1
numpy==1.24.2
opencv-python==4.6.0.66
paddleocr==2.6.1.3
paddlepaddle==2.4.2
Pillow==9.4.0
scikit-image==0.20.0
scipy==1.9.1
uvicorn==0.21.0
